- Rename package from uwpds_client to uw_person_client - Add CICD file - Add empty test file - Remove VERSION file
