+++ .github/workflows/cicd.yml
@@ -1,0 +1,96 @@
+#
+# Example GitHub Actions config for UW-IT AXD2 app testing and publishing to PyPi
+#
+# Preconditions:
+#
+#  1) Application contains a setup.py file
+#
+#  2) Application repo has access to the required secret
+#     at https://github.com/organizations/uw-it-aca/settings/secrets:
+#
+#     PYPI_API_TOKEN
+#
+# To adapt this config to a specific Python app:
+#
+#  1) Set APP_NAME to the name of the package name/directory.
+#
+#  2) Set CONF_PATH to a path containing a urls.py and/or settings.py file
+#     used by the test suite, if any
+#
+#  3) Verify that the lists of branches for push/pull_request is appropriate,
+#     and add other branch names if needed.
+#
+#  4) Update the matrix of django versions to test, if necessary.
+#
+---
+name: tests
+
+env:
+  APP_NAME: uw_person_client
+on:
+  push:
+    branches: [main, master, qa, develop]
+  pull_request:
+    branches: [main, master, qa, develop]
+    types: [opened, reopened, synchronize]
+  release:
+    branches: [main, master]
+    types: [published]
+
+jobs:
+  test:
+    runs-on: ubuntu-18.04
+
+    steps:
+      - name: Checkout Repo
+        uses: actions/checkout@v2
+
+      - name: Setup Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.6
+
+      - name: Run Python Linters
+        uses: uw-it-aca/actions/python-linters@main
+        with:
+          app_name: ${APP_NAME}
+
+      - name: Install Dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e .
+          pip install nose2 commonconf coverage coveralls==2.2.0
+      - name: Run Tests
+        env:
+          LIVE_TESTS: 1
+        run: |
+          python -m compileall ${APP_NAME}/
+          coverage run ${APP_NAME}/test.py -v
+      - name: Report Test Coverage
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        shell: bash
+        run: coveralls
+
+  publish:
+    if: github.event_name == 'release' && startsWith(github.ref, 'refs/tags')
+
+    needs: test
+
+    runs-on: ubuntu-18.04
+
+    steps:
+      - name: Checkout Repo
+        uses: actions/checkout@v2
+
+      - name: Setup Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: '3.x'
+
+      - name: Publish to PyPi
+        uses: uw-it-aca/actions/publish-pypi@main
+        with:
+          app_name: ${APP_NAME}
+          tag_name: ${{ github.event.release.tag_name }}
+          api_token: ${{ secrets.PYPI_API_TOKEN }}
