+++ README.md
@@ -1,6 +1,6 @@
-# uw-person-datastore-client
+# axdd-person-client
 
-[![Build Status](https://github.com/uw-it-aca/uw-person-datastore-client/workflows/Build%2C%20Test%20and%20Deploy/badge.svg?branch=main)](https://github.com/uw-it-aca/uw-person-datastore-client/actions)
-[![Coverage Status](https://coveralls.io/repos/github/uw-it-aca/uw-person-datastore-client/badge.svg?branch=main)](https://coveralls.io/github/uw-it-aca/uw-person-datastore?branch=main)
+[![Build Status](https://github.com/uw-it-aca/axdd-person-client/workflows/Build%2C%20Test%20and%20Deploy/badge.svg?branch=main)](https://github.com/uw-it-aca/axdd-person-client/actions)
+[![Coverage Status](https://coveralls.io/repos/github/uw-it-aca/axdd-person-client/badge.svg?branch=main)](https://coveralls.io/github/uw-it-aca/axdd-person-client?branch=main)
 
-A library for connecting to the UW Person Datastore.
+A client library for connecting to the UW Person Datastore.
