+++ uw_person_client/test.py
@@ -1,0 +1,3 @@
+# Copyright 2022 UW-IT, University of Washington
+# SPDX-License-Identifier: Apache-2.0
+
