--- uwpds_client/VERSION
@@ -1,1 +1,0 @@
-0.1
